updates to fasta files and chemidoc
in fasta files: {genomic region} to be replaced with the fasta output of the genome browser
---
chemidoc
maybe adding fields for specific files would help to keep track the relevant files for the experiment. add more values in chemidoc_protocol as needed for the images aquired. Also, (experiment) means that its is to be replaced with the experiment name (might help searches in the future)